[day-02] ayoub cpp solution fix

+++ day-02/part-1/ayoub.cpp
@@ -3,24 +3,24 @@
 
 using namespace std;
 
-uint64_t run(char* s) {
+int run(char* s) {
     size_t i = -1;
-    uint64_t curr = 0;
-    uint64_t h = 0, d = 0;
+    int curr = 0;
+    int h = 0, d = 0;
 
     while (s[++i]) {
         if (s[i] == '\n') continue;
         if (s[i] == 'f') {
             i += 7;
-            curr = 0; while (s[++i] != '\n' && s[i]) curr = curr*10ULL + (uint64_t)(s[i]-'0');
+            curr = 0; while (s[++i] != '\n' && s[i]) curr = curr*10 + (int)(s[i]-'0');
             h += curr;
         } else if (s[i] == 'u') {
             i += 2;
-            curr = 0; while (s[++i] != '\n' && s[i]) curr = curr*10ULL + (uint64_t)(s[i]-'0');
+            curr = 0; while (s[++i] != '\n' && s[i]) curr = curr*10 + (int)(s[i]-'0');
             d -= curr;
         } else if (s[i] == 'd') {
             i += 4;
-            curr = 0; while (s[++i] != '\n' && s[i]) curr = curr*10ULL + (uint64_t)(s[i]-'0');
+            curr = 0; while (s[++i] != '\n' && s[i]) curr = curr*10 + (int)(s[i]-'0');
             d += curr;
         }
     }
@@ -35,7 +35,7 @@ int main(int argc, char** argv) {
     }
 
     clock_t start = clock();
-    uint64_t answer = run(argv[1]);
+    int answer = run(argv[1]);
     
     cout << "_duration:" << float( clock () - start ) * 1000.0 /  CLOCKS_PER_SEC << "\n";
     cout << answer << "\n";
